Fixed issues with new fields and file errors

+++ dashboard/addItem.php
@@ -5,8 +5,8 @@ require_once "../models/Item.php";
 
 $error = '';
 if ($_SERVER['REQUEST_METHOD'] === 'POST') {
-    if (isset($_POST['itemName']) && isset($_POST['itemAuthor']) && isset($_POST['itemPrice']) && isset($_POST['itemStock']) && isset($_POST['itemDescription']) &&
-        !empty($_POST['itemName']) && !empty($_POST['itemAuthor']) && !empty($_POST['itemPrice']) && !empty($_POST['itemStock']) && !empty($_POST['itemDescription'])) {
+    if (isset($_POST['itemName']) && isset($_POST['itemAuthor']) && isset($_POST['itemPages']) && isset($_POST['itemPrice']) && isset($_POST['itemStock']) && isset($_POST['itemDescription']) &&
+        !empty($_POST['itemName']) && !empty($_POST['itemAuthor']) && !empty($_POST['itemPages']) && !empty($_POST['itemPrice']) && !empty($_POST['itemStock']) && !empty($_POST['itemDescription'])) {
         $item = new Item($conn);
         $item->seller_id = $_SESSION['id'];
         $item->name = $_POST['itemName'];
@@ -17,6 +17,7 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
                 $item->price = $_POST['itemPrice'];
                 $item->stock = $_POST['itemStock'];
                 $item->description = $_POST['itemDescription'];
+                $item->number_pages = $_POST['itemPages'];
 
                 $target_dir = "../data/product-images/";
                 $target_file = $target_dir . basename($_FILES["itemImage"]["name"]);
@@ -26,17 +27,21 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
                     $error = 'Image is too big';
                     if ($imageFileType != "jpg" && $imageFileType != "png" && $imageFileType != "jpeg") {
                         $error = "Sorry, only JPG, JPEG, PNG & GIF files are allowed.";
-                    } else if (move_uploaded_file($_FILES["itemImage"]["tmp_name"], $target_file)) {
-                        $item->image = $_FILES['itemImage']['name'];
-                        $item_id = $item->addItem();
-                        if ($item_id !== null) {
-                            header("Location: ../page.php?id=$item_id");
+                    } else if (strlen($_FILES['itemImage']['name']) <= 120) {
+                        if (move_uploaded_file($_FILES["itemImage"]["tmp_name"], $target_file)) {
+                            $item->image = $_FILES['itemImage']['name'];
+                            $item_id = $item->addItem();
+                            if ($item_id !== null) {
+                                header("Location: ../page.php?id=$item_id");
+                            } else {
+                                unlink($target_file);
+                                $error = "There was an error adding the item.";
+                            }
                         } else {
-                            unlink($target_file);
-                            $error = "There was an error adding the item.";
+                            $error = "Sorry, there was an error uploading your file.";
                         }
                     } else {
-                        $error = "Sorry, there was an error uploading your file.";
+                        $error = "File name is too long.";
                     }
                 } else {
                     $error = "File is not an image.";
@@ -77,6 +82,10 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
                 <div class="form-group">
                     <label for="itemAuthor">Author</label>
                     <input class="form-control" type="text" id="itemAuthor" name="itemAuthor" required/>
+                </div>
+                <div class="form-group">
+                    <label for="itemPages">Number of pages</label>
+                    <input class="form-control" type="text" id="itemPages" name="itemPages" required/>
                 </div>
                 <div class="form-group">
                     <label for="itemPrice">Price</label>
